Implemented one mechanism for handling spaces. Also, updated test cleanup code.
This might not be the best mechanism for handling extra space / characters
since you could wind up obliterating a comment on the same line as the code.
Probably the "best" thing to do is to find the actual header text in the line
and ensure you only replace that instead of rewriting the entire line. Its a
little more work than what I'm doing now though and I don't have a customer
that needs this (the currently implemented mechanism  works consistently on my
test codebase without sideeffects).