Turn on native testing and make them pass.
This enables testing of the native uh program and turns on two simple cases.
TestCase 1: header_filename which is a basic test to ensure that the tool is
    finding that header.
TestCase 2: testing whether or not it is picking up optional directories.

There is still a fair amount of work left to make the native tool match the
output of the python one as well as bringing over more of the python test
suite.